requirements.txt
streamlit>=1.31.0
google-generativeai>=0.3.2
pydantic>=2.5.2
python-dotenv>=1.0.1
pdfkit>=1.0.0
markdown>=3.5.2
wkhtmltopdf>=0.2 